fix failing parallel phold tests
When running any parallel phold_test*, it splits the communicator so any
uses of MPI_COMM_WORLD in ROSS can be caught. This increases the number
of ranks used in the parallel tests, so that when splitting the comm,
there is still more than one rank in the communicator used for the test.
It also has the rank that doesn't get passed to ROSS return after
MPI_Finalize so we don't get an error about calling MPI functions after
finalize.