wip improvements to cmake
Makes a lot of improvements to CMake, the biggest being that the
ROSSConfig.cmake is properly created and installed and targets are
correctly exported. This will make it easier to use find_package(ROSS)
by CMake in other libraries, instead of weirdly using pkgconfig through
CMake. The current pkgconfig setup is left as is and can be left until
CODES fully transitions to using CMake.